code refactor and refetch remove
